+++ Integral.h
@@ -87,7 +87,7 @@ class MonteCarlo: public Integral
         double operator()(double lower, double upper,
                           int n_points, Function *f);
         double get_variance();
-        virtual void set_seed(int seed){}
+        virtual void set_seed(long int seed){}
 };
 
 
@@ -99,7 +99,7 @@ class MonteCarloBF: public MonteCarlo
         MonteCarloBF(int dimension);
         double constant_term(double upper, double lower);
         double new_term(double lower, double upper);
-        void set_seed(int seed);
+        void set_seed(long int seed);
 };
 
 
@@ -111,7 +111,7 @@ class MonteCarloIS: public MonteCarlo
         MonteCarloIS(int dimension);
         double constant_term(double upper, double lower);
         double new_term(double upper, double lower);
-        void set_seed(int seed);
+        void set_seed(long int seed);
 };
 
 #endif // INTEGRAL_H
